wip: cleanup publisher code.
