+++ livox_ros_driver/launch/livox_lidar.launch
@@ -2,7 +2,7 @@
 
     <arg name="lvx_file_path" default="livox_test.lvx"/>
     <arg name="bd_list" default="100000000000000"/>
-    <arg name="xfer_format" default="0"/>
+    <arg name="xfer_format" default="2"/>
     <arg name="multi_topic" default="0"/>
     <arg name="data_src" default="0"/>
     <arg name="publish_freq" default="10.0"/>
