+++ livox_ros_driver/package.xml
@@ -27,6 +27,7 @@
     <depend>rospy</depend>
     <depend>sensor_msgs</depend>
     <depend>std_msgs</depend>
+    <depend>rosfmt</depend>
 
     <exec_depend>message_runtime</exec_depend>
     <exec_depend>rviz</exec_depend>
